requirements.txt
PyYAML==5.3.1
netmiko==3.1.0
netdev==0.9.3
